improve(MarkdownTextWrap): partial implementation of component
