+++ grapher/core/grapher.scss
@@ -137,3 +137,4 @@ $zindex-Tooltip: 20;
 @import "grapher/mapCharts/MapTooltip.scss";
 @import "grapher/sparkBars/SparkBars.scss";
 @import "grapher/loadingIndicator/LoadingIndicator.scss";
+@import "grapher/text/MarkdownTextWrap.scss";
