+++ grapher/text/MarkdownTextWrap.scss
@@ -1,0 +1,3 @@
+.markdown-text-wrap__line {
+    display: block;
+}
